refactor: stop exporting createSignature and add CLI script for generating signatures
