+++ src/index.ts
@@ -5,4 +5,4 @@ export {
   InvalidWebhookSignatureVersion,
 } from './signature-verifier.exceptions';
 export { SignatureVerificationResult } from './signature-verifier.types';
-export { createSignature } from './create-signature';
+
